compat: add memfd_create for BSD
FreeBSD uses shm_open with SHM_ANON (available since 8.0), while
OpenBSD and NetBSD just fail, so the fallback code could be used
instead.

linux checks for support in the kernel and if it is missing fails
(available since 3.17 and including debian 8)

most of the time the libc wrapper should be used instead since it
is available since glibc 2.27 and musl 1.1.20 and it is even
found in bionic (AKA android) with API 30, but for the cases where
libc is too old a syscall wrapper is provided.