protexec: enforce use of secure_getenv in create_tmpfile
use of getenv is not safe where this code could end up being executed
so only allow TMPDIR override when secure_getenv is available.

note that NetBSD >= 8 won't experience a regression as this code is
only used when there is no MAP_REMAPDUP and NetBSD < 8 shouldn't be
using this allocator as the default one works better.