protexec: restrict file system permissions in the map files
the backing file contents should not be accessible outside the running
process so make sure there are no RW permissions in the file created
for any user as early as possible.

memfd_create will create the file as 0600, and while it is not linked
anywhere in the filesystem is still accessible through its proc link
in /proc/$PID/fd/$FD; after this change that link won't be accessible
outside the running process.