wxexec: fail early if WX pages are forbidden per policy
in security enhanced systems like SELinux or PaX, the use of WX
pages might be forbidden by policy (unless exceptions are added).

since switching page permissions is assumed to work and does no
error checking, the result will be a segfault in the application.

instead check early and report the failure upstream by failing
the allocation.

note that unlike other allocators until now, the memory is not
returned in a writeable state, so add calls to flip from RX <-> RW
in all (except for TileGX) architectures generate_code().

reducing the window where the page is writeable is meant to increase
security.