wxexec: fail early if WX pages are forbidden per policy
in security enhanced systems like SELinux or PaX, the use of WX
pages might be forbidden by policy (unless exceptions are added).

since switching page permissions is assumed to work, the result
will be a segfault in the application.

check instead early and report the failure upstream by failing
the allocation.

note that unlike other allocators, the memory is not returned in
a writable state and therefore the corresponding generate_code
functions has been extended to make the memory writeable just
before the code is assembled.