macos: enable fastcall function calls
might had been accidentally disabled, because for a while,
macOS used to have an ancient (pre GPL3) gcc compiler that
might had lacked the necessary support.

that problem is likely covered with the recently added check,
and current "Apple LLVM" doesn't fail to build and seem to
work with it enabled.

tested with 10.3.6 (which is also the last version that supports
X86_32 bit applications) but could do with further testing.