execalloc: allow weak linking (WIP)
allows overriding the sljit_malloc_exec function at build time, but
only if sljitLir was enabled with weak symbol support by using
-DSLJIT_WEAK_EXECUTABLE_ALLOCATOR.

only useful as a hack for test64 that needs to be able to correctly
guess which address would be used if the labels were emitted.

the odd values for size are designed to force the compiler to set
a low address eventhough it will effectively wraparound.