x86: refactor sljit_label (WIP)
rename len to offset to better reflect what it means and the fact
that it could be negative (change type to match).

while at it make sure all interactions with the code ptr are done
as integers.

needs to be compile/tested in all other architectures and have TODO's
checked (integration with wextra and clean CI a plus).