x86: add bswap (POC)
will use SLJIT_BSWAP for the native word size unlike other compilers
that consider 32bit BSWAP the default even in 64bit[1].

SLJIT_BSWAP32 is provided to allow operations with the low half of
64 bit registers (and 0 expanding the higher part) to make the
results equivalent to 32 bit.

missing support for all other architectures and most importantly the
emulated fallback that was suggested.

also missing error checking (to avoid making it less didactic), and
has only a barebones test that should be expanded.

Closes: zherczeg/sljit#48

[1] https://www.felixcloutier.com/x86/bswap