debug: emulate apple silicon environment (NOMERGE)
use a hacked regular allocator to try to simulate apple silicon
by returning a non writablee page and use a helper function to
change all allocated pages permissions.

note that for it to work, a hacked version of a chunk array had
to be added as well, which is obviously not correct for all cases,
but correct enough to successful run the tests and do further
testing.

modify config so the code could be built and tested in POSIX
systems.