exec: add calls to WX where needed
since the allocator keeps the metadata about chunk/block in the
executable pages, then will need to enable/disable write access
to those pages also when metadata needs updating.

note that since there is no need for an accurate pointer range
with the apple implementation, NULL is passed for the pointers.

Fixes: zherczeg/sljit#51