wxexec: compatibility documentation update
clarify that the use of MAP_ANON is a hard requirement for cases where
it was not available and triggers a build failure (ex: really old *NIX
or strict POSIX).

while at it, update the wxexec code so it supports better really
ancient platforms that had a slightly different memory mapping API,
including some custom types (which is why we need an extra header).