macos: add BigSur support to execalloc
apple silicon requires that pages that will hold JIT code are
marked with MAP_JIT and that a call be made before writing to
them, so add support for both.

since the allocator keeps the metadata about chunk/block in the
executable pages, all functions that modify that metadata will
also need to be updated.

note that since there is no need for an accurate pointer range
with the apple implementation, NULL is passed for the pointers.

this path enables MAP_JIT also in BigSur Intel and uses the same
new interface that is required for Apple Silicon to aid testing,
but that could cause problems with fork() so it should only be
considered experimental.

Fixes: #51