macos: restrict MAP_JIT to Apple Silicon
MAP_JIT is a required flag for JIT enabled pages when the hardened
runtime is being used, but by doing so, adds several undocumented
restrictions that are not worth having when hardened runtime is not
enabled.

default BigSur to the original logic so MAP_JIT will only be enabled
when it would be useful, and while doing that restrict the additional
code to be only available for non x86 CPUs.