build: add barebones cmake for Windows
Building sljit requires GNU make, and uses a syntax that can be
easily converted into BSD or even NMAKE if needed, but in Windows
users could be tempted to try Msys or MinGW as a way to get make
working, only to find a broken library.

Provide instead with a cmake recipe that could be used to generate
whatever is that they use to build with native tools, and so be
able to test or improve sljit more easily.

This script is intentionally minimal to avoid duplicating the logic
from the makefile and even reads it to avoid hardcoding values that
are relevant to the build but to avoid overwriting that file, as
well as making it clear that it uses GNU syntax had it renamed.