export state of the executable allocator
Create an external interface to query the state of the executable
allocator, so that users of the library can report back at runtime
when JIT availability has been impaired by W^X policies.

Consolidate most checks to a single interface for easy of maintenance.

Support for !MAP_ANON and integration with the "wxexec allocator"
punted for now.