x86: avoid hardcoding rsi in execute_cpu_id
While it is likely to be the same assigned register, it is a
strange choice for the clobbered list as it is a saved register
in some ABIs.

Use a proper output operand and make sure to pass its pointer
as an input operand, but to make sure that is not assigned
to "ebx", use a local alias that is bound to "esi" in 32-bit.

Additionallu move the push/pop for ebx where it was before
this problem was introduced with e093197 (Improve Windows
compatiblity, 2023-11-01).

While at it, merge both almost identical branches to avoid
unnecessary code duplication.